+++ .editorconfig
@@ -10,10 +10,10 @@ root = true
 charset = utf-8
 
 # Indentation :  set to tab or space to use hard tabs or soft tabs.
-indent_style = tab
+indent_style = space
 
 # A whole number define	the number of column used for indentation
-indent_size = 4
+indent_size = 2
 
 #"true" esures file ends with a new line
 insert_final_newline = true
