+++ extra/index.js
@@ -1,6 +1,6 @@
 
-import {hasPermission} from './utils/permissions'
+import hasPermission from './extra/utils/permissions'
 
 
-console.log(hasPermission(MODULE_USER,ROLL_OF_HEAD_TRAINER,PERMISSION_DELETE));
-console.log(hasPermission(MODULE_USER,ROLE_OF_TRAINEE,PERMISSION_DELETE));
+console.log(hasPermission('getUsers','trainee','read'));
+
